feat: remove server-side content cache
The idea here is putting a CDN like Fastly in front of the app and
applying thoughtful Cache-Control headers obviates the need for a
server-side content cache.
