+++ .env.example
@@ -6,6 +6,5 @@ CONTENTFUL_ACCESS_TOKEN=
 CONTENTFUL_URL=
 CONTENTFUL_PREVIEW="true"
 CONTENTFUL_WEBHOOK_TOKEN=
-DATABASE_FILE_PATH="./app/db.server/cache.db"
 SITE_IMAGE_URL=
 SITE_IMAGE_ALT=
