+++ .gitignore
@@ -247,10 +247,6 @@ $RECYCLE.BIN/
 /build
 .env
 
-# Drizzle
-/app/db.server/*.db
-/app/db.server/*.db-*
-
 # Generated SVG icons sprite and correpsonding types
 /app/icons/icons-sprite.svg
 /app/icons/types.ts
