+++ Dockerfile
@@ -1,16 +1,12 @@
 # base node image
 FROM node:22-bullseye-slim AS base
 
-# Install fuse3 and ca-certificates for litefs
-RUN apt-get update && apt-get install -y fuse3 ca-certificates
-
 ENV NODE_ENV="production"
 
 # Install all node_modules, including dev dependencies
 FROM base AS deps
 
-RUN mkdir /app
-WORKDIR /app
+WORKDIR /myapp
 
 ADD package.json package-lock.json ./
 RUN npm install --include=dev
@@ -18,20 +14,18 @@ RUN npm install --include=dev
 # Setup production node_modules
 FROM base AS production-deps
 
-RUN mkdir /app
-WORKDIR /app
+WORKDIR /myapp
 
-COPY --from=deps /app/node_modules /app/node_modules
+COPY --from=deps /myapp/node_modules /myapp/node_modules
 ADD package.json package-lock.json ./
 RUN npm prune --omit=dev
 
 # Build the app
 FROM base AS build
 
-RUN mkdir /app
-WORKDIR /app
+WORKDIR /myapp
 
-COPY --from=deps /app/node_modules /app/node_modules
+COPY --from=deps /myapp/node_modules /myapp/node_modules
 
 ADD . .
 RUN npm run build
@@ -39,23 +33,14 @@ RUN npm run build
 # Finally, build the production image with minimal footprint
 FROM base
 
+ENV PORT="8080"
 ENV NODE_ENV="production"
 
-WORKDIR /app
+WORKDIR /myapp
 
-COPY --from=production-deps /app/node_modules /app/node_modules
-COPY --from=build /app/build/server /app/build/server
-COPY --from=build /app/build/client /app/build/client
+COPY --from=production-deps /myapp/node_modules /myapp/node_modules
+COPY --from=build /myapp/build /myapp/build
+COPY --from=build /myapp/public /myapp/public
+COPY --from=build /myapp/package.json /myapp/package.json
 
-# prepare litefs
-ENV LITEFS_DIR="/litefs"
-ENV DATABASE_FILE_PATH="$LITEFS_DIR/cache.db"
-ENV INTERNAL_PORT="8080"
-ENV PORT="8081"
-COPY --from=flyio/litefs:0.5 /usr/local/bin/litefs /usr/local/bin/litefs
-ADD litefs.yml /etc/litefs.yml
-RUN mkdir -p /data ${LITEFS_DIR}
-
-ADD . .
-
-ENTRYPOINT [ "litefs", "mount" ]
+ENTRYPOINT [ "npm", "run", "start" ]
