--- app/db.server/0000_thick_korvac.sql
@@ -1,14 +1,0 @@
--- Migrating from Prisma
-DROP TABLE IF EXISTS `_prisma_migrations`;
---> statement-breakpoint
-DROP TABLE IF EXISTS `CacheEntry`;
---> statement-breakpoint
-CREATE TABLE `cacheEntry` (
-	`id` integer PRIMARY KEY NOT NULL,
-	`createdAt` text DEFAULT CURRENT_TIMESTAMP NOT NULL,
-	`updatedAt` text NOT NULL,
-	`key` text NOT NULL,
-	`value` text NOT NULL
-);
---> statement-breakpoint
-CREATE UNIQUE INDEX `cacheEntry_key_unique` ON `cacheEntry` (`key`);
