--- app/db.server/meta/_journal.json
@@ -1,13 +1,0 @@
-{
-  "version": "5",
-  "dialect": "sqlite",
-  "entries": [
-    {
-      "idx": 0,
-      "version": "5",
-      "when": 1709622306297,
-      "tag": "0000_thick_korvac",
-      "breakpoints": true
-    }
-  ]
-}
