--- app/routes/_app.tags._index/loader.server.ts
@@ -1,55 +1,0 @@
-import { data, type LoaderFunctionArgs } from '@remix-run/node';
-import { getEnvVars } from '~/utils/env.server';
-import { fetchGraphQL } from '~/utils/graphql.server';
-import { cache } from '~/utils/cache.server';
-import type { Drink, DrinkTagsResponse } from '~/types';
-
-export async function loader({ request }: LoaderFunctionArgs) {
-  const cacheKey = new URL(request.url).pathname;
-  const cachedData: { tags: string[] } = await cache.get(cacheKey);
-  if (cachedData) return cachedData;
-
-  const { CONTENTFUL_ACCESS_TOKEN, CONTENTFUL_URL, CONTENTFUL_PREVIEW } = getEnvVars();
-
-  const allDrinkTagsQuery = /* GraphQL */ `
-    query ($preview: Boolean) {
-      drinkCollection(preview: $preview) {
-        drinks: items {
-          tags
-        }
-      }
-    }
-  `;
-
-  const queryResponse = await fetchGraphQL(
-    CONTENTFUL_URL,
-    CONTENTFUL_ACCESS_TOKEN,
-    allDrinkTagsQuery,
-    {
-      preview: CONTENTFUL_PREVIEW === 'true',
-    },
-  );
-
-  const queryResponseJson: DrinkTagsResponse = await queryResponse.json();
-
-  if (queryResponseJson.errors?.length || !queryResponseJson.data.drinkCollection) {
-    throw data(queryResponseJson, { status: 500 });
-  }
-
-  const {
-    data: {
-      drinkCollection: { drinks: maybeDrinks },
-    },
-  } = queryResponseJson;
-
-  const drinks = maybeDrinks.filter((drink): drink is Drink => Boolean(drink));
-  const uniqueTags = drinks.reduce<Set<string>>((acc, drink) => {
-    drink.tags?.forEach((tag) => acc.add(tag));
-    return acc;
-  }, new Set());
-
-  const loaderData = { tags: Array.from(uniqueTags).sort() };
-
-  await cache.set(cacheKey, loaderData);
-  return loaderData;
-}
