--- app/utils/prime-content-cache.server.ts
@@ -1,82 +1,0 @@
-import type { LoaderFunctionArgs } from '@remix-run/node';
-import { kebabCase } from 'lodash-es';
-import pThrottle from 'p-throttle';
-import { getEnvVars } from '~/utils/env.server';
-import { cache } from '~/utils/cache.server';
-import { loader as allDrinksLoader } from '~/routes/_app._index/loader.server';
-import { loader as drinkLoader } from '~/routes/_app.$slug/loader.server';
-import { loader as allTagsLoader } from '~/routes/_app.tags._index/loader.server';
-import { loader as tagLoader } from '~/routes/_app.tags.$tag/loader.server';
-
-const { CONTENTFUL_PREVIEW } = getEnvVars();
-
-// https://www.contentful.com/developers/docs/technical-limits/
-const throttle = pThrottle({
-  limit: CONTENTFUL_PREVIEW === 'true' ? 13 : 54,
-  interval: 1000,
-});
-
-export async function primeContentCache() {
-  try {
-    console.log('🚒 priming content cache...');
-
-    // 1. Start with an empty cache
-    await cache.clear();
-
-    // 2. Load and cache all drinks
-    const allDrinksDataFnArgs: LoaderFunctionArgs = {
-      context: {},
-      params: {},
-      request: new Request('https://drinks.fyi'),
-    };
-    const throttledAllDrinksLoader = throttle(allDrinksLoader);
-    const allDrinksData = await throttledAllDrinksLoader(allDrinksDataFnArgs);
-    const drinks = allDrinksData.drinks.filter(Boolean);
-
-    // 3. Load and cache each individual drink
-    const allSlugs = drinks.map(({ slug }) => slug);
-    const throttledDrinkLoader = throttle(drinkLoader);
-    await Promise.all(
-      allSlugs.map(async (slug) => {
-        const drinkDataFnArgs: LoaderFunctionArgs = {
-          context: {},
-          params: { slug },
-          request: new Request(`https://drinks.fyi/${slug}`),
-        };
-        return throttledDrinkLoader(drinkDataFnArgs);
-      }),
-    );
-
-    // 4. Load and cache all tags
-    const allTagsDataFnArgs: LoaderFunctionArgs = {
-      context: {},
-      params: {},
-      request: new Request('https://drinks.fyi/tags'),
-    };
-    const throttledAllTagsLoader = throttle(allTagsLoader);
-    const allTagsData = await throttledAllTagsLoader(allTagsDataFnArgs);
-    const { tags } = allTagsData;
-
-    // 5. Load and cache each individual tag
-    const throttledTagLoader = throttle(tagLoader);
-    await Promise.all(
-      tags.map(async (tag) => {
-        const tagSlug = kebabCase(tag);
-        const tagDataFnArgs: LoaderFunctionArgs = {
-          context: {},
-          params: { tag: tagSlug },
-          request: new Request(`https://drinks.fyi/tags/${tagSlug}`),
-        };
-        return throttledTagLoader(tagDataFnArgs);
-      }),
-    );
-
-    console.log('🔥 content cache is warm');
-  } catch (err) {
-    if (err instanceof Response) {
-      console.error(`💣 failed to warm cache: ${await err.text()}`);
-    } else {
-      console.error(err);
-    }
-  }
-}
