+++ fly.toml
@@ -9,11 +9,6 @@ kill_signal = "SIGINT"
 kill_timeout = "5s"
 swap_size_mb = 512
 
-[mounts]
-  source = "data"
-  destination = "/data"
-  processes = ["app"]
-
 [[services]]
   protocol = "tcp"
   internal_port = 8080
