+++ package.json
@@ -27,7 +27,6 @@
     "format": "prettier --cache --write .",
     "format:check": "prettier --cache --check .",
     "lint": "eslint --cache --cache-location ./node_modules/.cache/eslint .",
-    "setup": "tsx ./app/db.server/migrate.ts",
     "start": "cross-env NODE_ENV=production tsx ./server.ts",
     "typecheck": "tsc",
     "validate": "run-p format:check lint typecheck"
@@ -42,18 +41,13 @@
     "@remix-run/react": "2.13.1",
     "@remix-run/serve": "2.13.1",
     "algoliasearch": "5.12.0",
-    "better-sqlite3": "11.5.0",
     "clsx": "2.1.1",
     "cross-env": "7.0.3",
-    "drizzle-orm": "0.36.0",
     "express": "4.21.1",
     "isbot": "5.1.17",
-    "litefs-js": "1.1.2",
     "lodash-es": "4.17.21",
     "marked": "14.1.3",
     "morgan": "1.10.0",
-    "p-throttle": "6.2.0",
-    "pathe": "1.1.2",
     "pretty-cache-header": "1.0.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
@@ -66,7 +60,6 @@
     "@remix-pwa/dev": "3.1.0",
     "@remix-run/dev": "2.13.1",
     "@sly-cli/sly": "1.14.0",
-    "@types/better-sqlite3": "7.6.11",
     "@types/express": "4.17.21",
     "@types/lodash-es": "4.17.12",
     "@types/morgan": "1.9.9",
@@ -77,7 +70,6 @@
     "@wkovacs64/prettier-config": "4.1.1",
     "autoprefixer": "10.4.20",
     "del-cli": "6.0.0",
-    "drizzle-kit": "0.27.1",
     "eslint": "8.57.1",
     "glob": "11.0.0",
     "node-html-parser": "6.1.13",
@@ -86,7 +78,6 @@
     "prettier": "3.3.3",
     "prettier-plugin-tailwindcss": "0.6.8",
     "semantic-release": "24.2.0",
-    "shelljs": "0.8.5",
     "tailwindcss": "3.4.14",
     "typescript": "5.6.3",
     "vite": "5.4.10",
